requirements.txt
Flask==3.1.3
Pillow==10.2.0
Werkzeug==3.1.6
gunicorn==21.2.0
